create Episode model
